+++ Login/Stage2.Designer.cs
@@ -32,12 +32,13 @@
             this.codeLabel = new System.Windows.Forms.Label();
             this.codeText = new System.Windows.Forms.TextBox();
             this.label1 = new System.Windows.Forms.Label();
+            this.errorLabel = new System.Windows.Forms.Label();
             this.SuspendLayout();
             // 
             // confirmButton
             // 
-            this.confirmButton.Location = new System.Drawing.Point(93, 202);
-            this.confirmButton.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.confirmButton.Location = new System.Drawing.Point(108, 90);
+            this.confirmButton.Margin = new System.Windows.Forms.Padding(2);
             this.confirmButton.Name = "confirmButton";
             this.confirmButton.Size = new System.Drawing.Size(75, 23);
             this.confirmButton.TabIndex = 0;
@@ -48,41 +49,51 @@
             // codeLabel
             // 
             this.codeLabel.AutoSize = true;
-            this.codeLabel.Location = new System.Drawing.Point(21, 159);
+            this.codeLabel.Location = new System.Drawing.Point(57, 68);
+            this.codeLabel.Margin = new System.Windows.Forms.Padding(2, 0, 2, 0);
             this.codeLabel.Name = "codeLabel";
-            this.codeLabel.Size = new System.Drawing.Size(122, 17);
+            this.codeLabel.Size = new System.Drawing.Size(92, 13);
             this.codeLabel.TabIndex = 1;
             this.codeLabel.Text = "Confirmation code";
             // 
             // codeText
             // 
-            this.codeText.Location = new System.Drawing.Point(149, 156);
-            this.codeText.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.codeText.Location = new System.Drawing.Point(153, 66);
+            this.codeText.Margin = new System.Windows.Forms.Padding(2);
             this.codeText.Name = "codeText";
-            this.codeText.Size = new System.Drawing.Size(100, 22);
+            this.codeText.Size = new System.Drawing.Size(76, 20);
             this.codeText.TabIndex = 2;
             // 
             // label1
             // 
-            this.label1.Location = new System.Drawing.Point(25, 37);
-            this.label1.Margin = new System.Windows.Forms.Padding(4, 0, 4, 0);
+            this.label1.Location = new System.Drawing.Point(12, 9);
             this.label1.Name = "label1";
-            this.label1.Size = new System.Drawing.Size(225, 84);
+            this.label1.Size = new System.Drawing.Size(260, 55);
             this.label1.TabIndex = 3;
             this.label1.Text = "You should now receive a confirmation code to the email address you entered in th" +
     "e last step.  Enter that code here and hit \'Confirm\' to finish the process";
             this.label1.TextAlign = System.Drawing.ContentAlignment.MiddleLeft;
             // 
+            // errorLabel
+            // 
+            this.errorLabel.ForeColor = System.Drawing.SystemColors.MenuHighlight;
+            this.errorLabel.Location = new System.Drawing.Point(15, 128);
+            this.errorLabel.Name = "errorLabel";
+            this.errorLabel.Size = new System.Drawing.Size(257, 124);
+            this.errorLabel.TabIndex = 4;
+            this.errorLabel.TextAlign = System.Drawing.ContentAlignment.TopCenter;
+            // 
             // Stage2
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(283, 254);
-            this.Controls.Add(this.label1);
+            this.ClientSize = new System.Drawing.Size(284, 261);
+            this.Controls.Add(this.errorLabel);
             this.Controls.Add(this.codeText);
             this.Controls.Add(this.codeLabel);
             this.Controls.Add(this.confirmButton);
-            this.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.Controls.Add(this.label1);
+            this.Margin = new System.Windows.Forms.Padding(2);
             this.Name = "Stage2";
             this.Text = "Elite: Dangerous login";
             this.ResumeLayout(false);
@@ -96,5 +107,6 @@
         private System.Windows.Forms.Label codeLabel;
         private System.Windows.Forms.TextBox codeText;
         private System.Windows.Forms.Label label1;
+        private System.Windows.Forms.Label errorLabel;
     }
 }